// hypixel-stats/hypixel-stats/types/hypixel.ts
export interface DuelsStats {
  wins?: number;
  losses?: number;
  kills?: number;
  deaths?: number;
  games_played_duels?: number;
  winstreak?: number;
  coins?: number;
  bow_shots?: number;
  bow_hits?: number;
  melee_swings?: number;
  melee_hits?: number;
}